♻️ : shared/ui Card 분리
